adicionar a controler do delete sale no module e adicionado a verificação de existência!

+++ src/usecases/sale/delete-sale/delete-sale.service.ts
@@ -1,10 +1,15 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { IsaleItemRepositoy } from 'src/domain/repository/ISaleItemRepository';
+import { messages } from 'src/util/error-messages';
 
 @Injectable()
 export class DeleteSaleService {
   constructor(private readonly repo: IsaleItemRepositoy) {}
   async execute(saleId: string) {
+    const ifSaleExists = await this.repo.findOne(saleId);
+
+    if (!ifSaleExists) throw new NotFoundException(messages.saleNotFound);
+
     return this.repo.delete(saleId);
   }
 }
